Refactor utils
- Refactor utils.py into ultils module
- Split into sectioned files for organisation and clarity
