+++ src/r0tools_simple_toolbox/utils/context.py
@@ -1,0 +1,94 @@
+from pathlib import Path
+
+import bpy
+
+from r0tools_simple_toolbox.const import INTERNAL_NAME
+
+
+def get_scene() -> bpy.types.Scene:
+    """Get the current scene"""
+    return bpy.context.scene
+
+
+def get_scene_name() -> str:
+    """Get the name of the current scene"""
+    return get_scene().name
+
+
+def get_context_area() -> str | None:
+    """Get the current area type or None if unavailable"""
+    if not bpy.context.area:
+        return None
+    return bpy.context.area.ui_type
+
+
+def get_addon_props():
+    """Get the addon property group from current scene"""
+    return get_scene().r0fl_toolbox_props
+
+
+def get_addon_prefs():
+    """Get the addon preferences"""
+    return bpy.context.preferences.addons[INTERNAL_NAME].preferences
+
+
+def get_addon_fs_path() -> Path:
+    """Get the filesystem path to the addon directory from THIS file"""
+    return Path(__file__).resolve().parent.parent
+
+
+def get_depsgraph():
+    """Get the current dependency graph"""
+    return bpy.context.evaluated_depsgraph_get()
+
+
+def get_depsgraph_is_updated_geometry() -> bool:
+    """Check if geometry was updated in the depsgraph"""
+    d = get_depsgraph()
+    for update in d.updates:
+        return update.is_updated_geometry
+    return False
+
+
+def get_depsgraph_is_updated_shading() -> bool:
+    """Check if shading was updated in the depsgraph"""
+    d = get_depsgraph()
+    for update in d.updates:
+        return update.is_updated_shading
+    return False
+
+
+def get_depsgraph_is_updated_transform() -> bool:
+    """Check if transforms were updated in the depsgraph"""
+    d = get_depsgraph()
+    for update in d.updates:
+        return update.is_updated_transform
+    return False
+
+
+def get_uvmap_size_x():
+    """Get selected UV Map Size in X"""
+    addon_props = get_addon_props()
+    return int(addon_props.uv_size_x)
+
+
+def get_uvmap_size_y():
+    """Get selected UV Map Size in Y"""
+    addon_props = get_addon_props()
+    return int(addon_props.uv_size_y)
+
+
+def save_preferences():
+    """Safely save user preferences without causing recursion"""
+    try:
+        if not hasattr(save_preferences, "is_saving"):
+            save_preferences.is_saving = False
+
+        if not save_preferences.is_saving:
+            save_preferences.is_saving = True
+            bpy.context.preferences.use_preferences_save = True
+            bpy.ops.wm.save_userpref()
+            save_preferences.is_saving = False
+    except Exception as e:
+        print(f"Error saving preferences: {e}")
+        save_preferences.is_saving = False
