+++ src/r0tools_simple_toolbox/utils/object_sets.py
@@ -1,0 +1,96 @@
+from .context import get_addon_prefs, get_addon_props
+from .general import IS_DEBUG
+from .handlers import queue_op
+
+
+def draw_objects_sets_uilist(layout, context, object_sets_box=None):
+    """
+    Draw the Objects Sets UI list
+
+    Args:
+        layout: The layout to draw in
+        context: The current context
+        object_sets_box: Optional box to draw within
+    """
+    addon_prefs = get_addon_prefs()
+    addon_props = get_addon_props()
+
+    # Object Sets Editor parent layout
+    if object_sets_box:
+        parent = object_sets_box
+    elif layout:
+        parent = layout
+    else:
+        print(f"[ERROR] No valid layout to use:\n{layout=}\n{object_sets_box=}")
+        return False
+
+    # Object Sets Use Colour
+    row = parent.row()
+    row.prop(addon_prefs, "object_sets_use_colour")
+
+    # Object Sets Row Number Slider (Same as in addon preferences)
+    row = parent.row()
+    row.prop(addon_prefs, "object_sets_list_rows", text="Rows:")
+
+    row = parent.row()
+    split = row.split(factor=0.92)  # Affects right side button width
+
+    # Left Section - List
+    col = split.column()
+    col.template_list(
+        "R0PROP_UL_ObjectSetsList",
+        "object_sets",
+        addon_props,  # Collection owner
+        "object_sets",  # Collection property
+        addon_props,  # Active item owner
+        "object_sets_index",  # Active item property
+        rows=addon_prefs.object_sets_list_rows,
+    )
+
+    # Right side - Buttons
+    col = split.column(align=True)
+    col.operator("r0tools.add_object_set_popup", text="+")
+    col.operator("r0tools.remove_object_set", text="-")
+    if len(addon_props.object_sets) > 1:  # Show buttons only when applicable
+        col.separator(factor=1.0)  # Spacer
+        col.operator("r0tools.move_object_set_item_up", icon="TRIA_UP", text="")
+        col.operator("r0tools.move_object_set_item_down", icon="TRIA_DOWN", text="")
+
+    col.separator(factor=1.0)  # Spacer
+    col.operator("r0tools.object_sets_refresh", text="", icon="FILE_REFRESH")
+
+    col.separator(factor=1.0)  # Spacer
+    col.operator("r0tools.object_sets_colours_randomise", text="", icon="NODE_MATERIAL")
+
+    # Bottom
+    if object_sets_box:
+        parent = object_sets_box
+    else:
+        parent = layout
+    row = parent.row(align=True)
+
+    # Add/Remove Object Set Buttons
+    split = row.split(factor=0.65)
+    row_col = split.row(align=True)
+    row_col.operator("r0tools.assign_to_object_set")
+    row_col.operator("r0tools.remove_from_object_set")
+    # Select Object Set Button
+    row_col = split.row()
+    op = row_col.operator("r0tools.select_object_set")
+    op.set_index = -1
+
+
+def refresh_object_sets_colours(context):
+    """Refresh colors for all object sets"""
+    print("[INFO] Force Refreshing Object Sets")
+    addon_prefs = get_addon_prefs()
+    addon_props = get_addon_props()
+    object_sets = addon_props.object_sets
+
+    if not addon_prefs.object_sets_use_colour:
+        return
+
+    for object_set in object_sets:
+        if IS_DEBUG():
+            print(f"[DEBUG] Refresh: {object_set.name}")
+        queue_op(object_set.update_object_set_colour, context)
